+++ guillotina/api/addons.py
@@ -1,10 +1,10 @@
 from guillotina import configure
 from guillotina import error_reasons
 from guillotina._settings import app_settings
-from guillotina.browser import ErrorResponse
 from guillotina.i18n import MessageFactory
 from guillotina.interfaces import IAddons
 from guillotina.interfaces import IContainer
+from guillotina.response import ErrorResponse
 from guillotina.utils import apply_coroutine