+++ guillotina/test_package.py
@@ -1,6 +1,4 @@
 # this is for testing.py, do not import into other modules
-from guillotina.interfaces import IObjectAddedEvent
-from aiohttp.web_exceptions import HTTPUnprocessableEntity
 from guillotina import configure
 from guillotina import schema
 from guillotina.async_util import IAsyncUtility
@@ -14,7 +12,9 @@ from guillotina.files import CloudFileField
 from guillotina.interfaces import IApplication
 from guillotina.interfaces import IContainer
 from guillotina.interfaces import IItem
+from guillotina.interfaces import IObjectAddedEvent
 from guillotina.interfaces import IResource
+from guillotina.response import HTTPUnprocessableEntity
 from zope.interface import implementer
 from zope.interface import Interface