+++ guillotina/traversal.py
@@ -1,59 +1,57 @@
 """Main routing traversal class."""
 from aiohttp.abc import AbstractMatchInfo
 from aiohttp.abc import AbstractRouter
-from aiohttp.web_exceptions import HTTPBadRequest
-from aiohttp.web_exceptions import HTTPException
-from aiohttp.web_exceptions import HTTPMethodNotAllowed
-from aiohttp.web_exceptions import HTTPNotFound
-from aiohttp.web_exceptions import HTTPUnauthorized
 from contextlib import contextmanager
 from guillotina import __version__
+from guillotina import error_reasons
 from guillotina import logger
+from guillotina import response
 from guillotina import routes
 from guillotina._settings import app_settings
 from guillotina.api.content import DefaultOPTIONS
 from guillotina.auth.participation import AnonymousParticipation
-from guillotina.browser import ErrorResponse
-from guillotina.browser import Response
-from guillotina.browser import UnauthorizedResponse
 from guillotina.component import get_adapter
 from guillotina.component import get_utility
 from guillotina.component import query_adapter
 from guillotina.component import query_multi_adapter
-from guillotina.contentnegotiation import content_type_negotiation
-from guillotina.contentnegotiation import language_negotiation
+from guillotina.contentnegotiation import get_acceptable_content_types
+from guillotina.contentnegotiation import get_acceptable_languages
 from guillotina.event import notify
 from guillotina.events import ObjectLoadedEvent
 from guillotina.exceptions import ConflictError
 from guillotina.exceptions import TIDConflictError
-from guillotina.exceptions import Unauthorized
 from guillotina.i18n import default_message_factory as _
 from guillotina.interfaces import ACTIVE_LAYERS_KEY
+from guillotina.interfaces import IAioHTTPResponse
 from guillotina.interfaces import IAnnotations
+from guillotina.browser import View
 from guillotina.interfaces import IApplication
 from guillotina.interfaces import IAsyncContainer
 from guillotina.interfaces import IContainer
 from guillotina.interfaces import IDatabase
 from guillotina.interfaces import IErrorResponseException
 from guillotina.interfaces import IInteraction
-from guillotina.interfaces import IOPTIONS
+from guillotina.interfaces import IOPTIONS, ILanguage
 from guillotina.interfaces import IParticipation
 from guillotina.interfaces import IPermission
-from guillotina.interfaces import IRendered
+from guillotina.interfaces import IRenderer
 from guillotina.interfaces import IRequest
 from guillotina.interfaces import IResource
-from guillotina.interfaces import ITranslated
 from guillotina.interfaces import ITraversable
 from guillotina.profile import profilable
 from guillotina.registry import REGISTRY_DATA_KEY
+from guillotina.response import HTTPBadRequest
+from guillotina.response import HTTPMethodNotAllowed
+from guillotina.response import HTTPNotFound
+from guillotina.response import HTTPUnauthorized
 from guillotina.security.utils import get_view_permission
 from guillotina.transactions import abort
 from guillotina.transactions import commit
 from guillotina.utils import import_class
 from zope.interface import alsoProvides
 
+import aiohttp
 import traceback
-import ujson
 import uuid
 
 
@@ -111,14 +109,6 @@ async def traverse(request, parent, path):
     return await traverse(request, context, path[1:])
 
 
-def generate_unauthorized_response(e, request):
-    # We may need to check the roles of the users to show the real error
-    eid = uuid.uuid4().hex
-    message = _('Not authorized to render operation') + ' ' + eid
-    logger.error(message, exc_info=e, eid=eid, request=request)
-    return UnauthorizedResponse(message, eid=eid)
-
-
 def generate_error_response(e, request, error, status=500):
     # We may need to check the roles of the users to show the real error
     eid = uuid.uuid4().hex
@@ -131,7 +121,12 @@ def generate_error_response(e, request, error, status=500):
         return http_response
     message = _('Error on execution of view') + ' ' + eid
     logger.error(message, exc_info=e, eid=eid, request=request)
-    return ErrorResponse(error, message, status=status, eid=eid)
+    return response.HTTPInternalServerError(content={
+        'message': message,
+        'reason': error_reasons.UNKNOWN.name,
+        'details': error_reasons.UNKNOWN.details,
+        'eid': eid
+    })
 
 
 class BaseMatchInfo(AbstractMatchInfo):
@@ -209,15 +204,27 @@ class BaseMatchInfo(AbstractMatchInfo):
                 resp.headers['XG-Error'] = 'Could not get stats'
 
 
+async def apply_rendering(view, request, view_result):
+    for ct in get_acceptable_content_types(request):
+        renderer = query_multi_adapter(
+            (view, request), IRenderer, name=ct)
+        if renderer is not None:
+            break
+    else:
+        # default to application/json
+        renderer = query_multi_adapter(
+            (view, request), IRenderer, name='application/json')
+    return await renderer(view_result)
+
+
 class MatchInfo(BaseMatchInfo):
     """Function that returns from traversal request on aiohttp."""
 
-    def __init__(self, resource, request, view, rendered):
+    def __init__(self, resource, request, view):
         super().__init__()
         self.request = request
         self.resource = resource
         self.view = view
-        self.rendered = rendered
 
     @profilable
     async def handler(self, request):
@@ -229,23 +236,11 @@ class MatchInfo(BaseMatchInfo):
                 # We try to avoid collisions on the same instance of
                 # guillotina
                 view_result = await self.view()
-                if isinstance(view_result, ErrorResponse) or \
-                        isinstance(view_result, UnauthorizedResponse):
-                    # If we don't throw an exception and return an specific
-                    # ErrorReponse just abort
-                    await abort(request)
-                    request._view_error = True
-                else:
-                    await commit(request, warn=False)
-
-            except Unauthorized as e:
-                await abort(request)
-                view_result = generate_unauthorized_response(e, request)
-                request._view_error = True
+                await commit(request, warn=False)
             except (ConflictError, TIDConflictError) as e:
                 # bubble this error up
                 raise
-            except HTTPException as exc:
+            except (response.Response, aiohttp.web_exceptions.HTTPException) as exc:
                 await abort(request)
                 view_result = exc
                 request._view_error = True
@@ -257,10 +252,7 @@ class MatchInfo(BaseMatchInfo):
         else:
             try:
                 view_result = await self.view()
-            except Unauthorized as e:
-                request._view_error = True
-                view_result = generate_unauthorized_response(e, request)
-            except HTTPException as exc:
+            except (response.Response, aiohttp.web_exceptions.HTTPException) as exc:
                 view_result = exc
                 request._view_error = True
             except Exception as e:
@@ -270,26 +262,24 @@ class MatchInfo(BaseMatchInfo):
                 await abort(request)
         request.record('viewrendered')
 
-        # Make sure its a Response object to send to renderer
-        if not isinstance(view_result, Response):
-            view_result = Response(view_result)
-        elif view_result is None:
-            # Always provide some response to work with
-            view_result = Response({})
+        if IAioHTTPResponse.providedBy(view_result):
+            resp = view_result
+        else:
+            resp = await apply_rendering(self.view, self.request, view_result)
+            request.record('renderer')
 
-        # Apply cors if its needed
-        cors_renderer = app_settings['cors_renderer'](request)
-        cors_headers = await cors_renderer.get_headers()
-        cors_headers.update(view_result.headers)
-        view_result.headers = cors_headers
-        retry_attempts = getattr(request, '_retry_attempt', 0)
-        if retry_attempts > 0:
-            view_result.headers['X-Retry-Transaction-Count'] = str(retry_attempts)
-
-        request.record('headers')
-
-        resp = await self.rendered(view_result)
-        request.record('rendered')
+            # Apply cors if its needed
+            cors_renderer = app_settings['cors_renderer'](request)
+            try:
+                cors_headers = await cors_renderer.get_headers()
+                cors_headers.update(resp.headers)
+                resp._headers = cors_headers
+                retry_attempts = getattr(request, '_retry_attempt', 0)
+                if retry_attempts > 0:
+                    resp.headers['X-Retry-Transaction-Count'] = str(retry_attempts)
+            except response.Response as exc:
+                resp = exc
+            request.record('headers')
 
         if not request._view_error:
             request.execute_futures()
@@ -325,7 +315,11 @@ class BasicMatchInfo(BaseMatchInfo):
         """Main handler function for aiohttp."""
         request.record('finish')
         self.debug(request, self.resp)
-        return self.resp
+        if IAioHTTPResponse.providedBy(self.resp):
+            return self.resp
+        else:
+            return await apply_rendering(
+                View(None, request), request, self.resp)
 
     def get_info(self):
         return {
@@ -355,18 +349,21 @@ class TraversalRouter(AbstractRouter):
         result = None
         try:
             result = await self.real_resolve(request)
-        except HTTPException as exc:
+        except (response.Response, aiohttp.web_exceptions.HTTPException) as exc:
             await abort(request)
             return BasicMatchInfo(request, exc)
-        except Exception as e:
-            logger.error("Exception on resolve execution", exc_info=e, request=request)
+        except Exception:
+            logger.error("Exception on resolve execution",
+                         exc_info=True, request=request)
             await abort(request)
-            raise e
+            return BasicMatchInfo(
+                request, response.HTTPInternalServerError())
+
         if result is not None:
             return result
         else:
             await abort(request)
-            return BasicMatchInfo(request, HTTPNotFound())
+            return BasicMatchInfo(request, response.HTTPNotFound())
 
     @profilable
     async def real_resolve(self, request: IRequest) -> MatchInfo:
@@ -377,9 +374,6 @@ class TraversalRouter(AbstractRouter):
             raise HTTPMethodNotAllowed(
                 method=request.method, allowed_methods=[k for k in app_settings['http_methods']])
         method = app_settings['http_methods'][request.method]
-
-        language = language_negotiation(request)
-        language_object = language(request)
 
         try:
             resource, tail = await self.traverse(request)
@@ -397,7 +391,9 @@ class TraversalRouter(AbstractRouter):
             }
             if app_settings.get('debug'):
                 data['traceback'] = traceback.format_exc()
-            raise HTTPBadRequest(text=ujson.dumps(data))
+            raise HTTPBadRequest(content={
+                'reason': data
+            })
 
         request.record('traversed')
 
@@ -420,10 +416,12 @@ class TraversalRouter(AbstractRouter):
         await self.apply_authorization(request)
         request.record('authentication')
 
-        translator = query_adapter(language_object, ITranslated,
-                                   args=[resource, request])
-        if translator is not None:
-            resource = translator.translate()
+        for language in get_acceptable_languages(request):
+            translator = query_adapter((resource, request), ILanguage,
+                                       name=language)
+            if translator is not None:
+                resource = translator.translate()
+                break
 
         # Add anonymous participation
         if len(security.participations) == 0:
@@ -453,7 +451,14 @@ class TraversalRouter(AbstractRouter):
                             auths=str([x.principal.id
                                        for x in security.participations])),
                         request=request)
-                    raise HTTPUnauthorized()
+                    raise HTTPUnauthorized(
+                        content={
+                            "reason": "You are not authorized to access content",
+                            "content": str(resource),
+                            "auths": [x.principal.id
+                                      for x in security.participations]
+                        }
+                    )
 
         if not view and len(tail) > 0:
             # we should have a view in this case because we are matching routes
@@ -468,10 +473,12 @@ class TraversalRouter(AbstractRouter):
         if not security.check_permission(view_permission, view):
             if IOPTIONS != method:
                 raise HTTPUnauthorized(
-                    reason="You are not authorized to view {content} with {auths}".format(
-                        content=resource,
-                        auths=str([x.principal.id
-                                   for x in security.participations]))
+                    content={
+                        "reason": "You are not authorized to view",
+                        "content": str(resource),
+                        "auths": [x.principal.id
+                                  for x in security.participations]
+                    }
                 )
 
         try:
@@ -486,18 +493,7 @@ class TraversalRouter(AbstractRouter):
 
         request.record('authorization')
 
-        renderer = content_type_negotiation(request, resource, view)
-        renderer_object = renderer(request)
-
-        rendered = query_multi_adapter(
-            (renderer_object, view, request), IRendered)
-
-        request.record('renderer')
-
-        if rendered is not None:
-            return MatchInfo(resource, request, view, rendered)
-        else:
-            return None
+        return MatchInfo(resource, request, view)
 
     async def traverse(self, request: IRequest) -> IResource:
         """Wrapper that looks for the path based on aiohttp API."""